requirements.txt
fastapi
uvicorn
requests
feedparser
google-generativeai
python-dotenv